Safari-609.1.11.1/LayoutTests/fast/frames/sandboxed-iframe-forms-dynamic-expected.txt

CONSOLE MESSAGE: line 3: Blocked form submission to 'javascript:window.top.disallowedFormSubmitted();' because the form's frame is sandboxed and the 'allow-forms' permission is not set.
    
PASS
